+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/AuditAdapter.java
@@ -22,11 +22,14 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
+import com.google.common.annotations.VisibleForTesting;
+
 import com.ericsson.bss.cassandra.ecaudit.common.record.SimpleAuditOperation;
 import com.ericsson.bss.cassandra.ecaudit.common.record.Status;
 import com.ericsson.bss.cassandra.ecaudit.entry.AuditEntry;
 import com.ericsson.bss.cassandra.ecaudit.entry.PreparedAuditOperation;
 import com.ericsson.bss.cassandra.ecaudit.entry.factory.AuditEntryBuilderFactory;
+import com.ericsson.bss.cassandra.ecaudit.entry.suppressor.BoundValueSuppressor;
 import com.ericsson.bss.cassandra.ecaudit.facade.Auditor;
 import com.ericsson.bss.cassandra.ecaudit.utils.Exceptions;
 import org.apache.cassandra.cql3.BatchQueryOptions;
@@ -49,6 +52,7 @@ public class AuditAdapter
 
     private final Auditor auditor;
     private final AuditEntryBuilderFactory entryBuilderFactory;
+    private BoundValueSuppressor boundValueSuppressor;
 
     private final Map<MD5Digest, String> idQueryCache = new ConcurrentHashMap<>();
 
@@ -57,11 +61,13 @@ public class AuditAdapter
      *
      * @param auditor             the auditor to use
      * @param entryBuilderFactory the audit entry builder factory to use
+     * @param boundValueSuppressor the bound value suppressor
      */
-    AuditAdapter(Auditor auditor, AuditEntryBuilderFactory entryBuilderFactory)
+    AuditAdapter(Auditor auditor, AuditEntryBuilderFactory entryBuilderFactory, BoundValueSuppressor boundValueSuppressor)
     {
         this.auditor = auditor;
         this.entryBuilderFactory = entryBuilderFactory;
+        this.boundValueSuppressor = boundValueSuppressor;
     }
 
     public static AuditAdapter getInstance()
@@ -122,7 +128,7 @@ public class AuditAdapter
                                                      .client(state.getRemoteAddress())
                                                      .coordinator(FBUtilities.getBroadcastAddress())
                                                      .user(state.getUser().getName())
-                                                     .operation(new PreparedAuditOperation(idQueryCache.get(id), options))
+                                                     .operation(new PreparedAuditOperation(idQueryCache.get(id), options, boundValueSuppressor))
                                                      .status(status)
                                                      .timestamp(timestamp)
                                                      .build();
@@ -234,7 +240,7 @@ public class AuditAdapter
             if (queryOrId instanceof MD5Digest)
             {
                 entryBuilderFactory.updateBatchEntryBuilder(builder, batchStatement.getStatements().get(statementIndex));
-                builder.operation(new PreparedAuditOperation(idQueryCache.get(queryOrId), options.forStatement(statementIndex)));
+                builder.operation(new PreparedAuditOperation(idQueryCache.get(queryOrId), options.forStatement(statementIndex), boundValueSuppressor));
                 batchOperations.add(builder.build());
             }
             else
@@ -252,5 +258,17 @@ public class AuditAdapter
     public Auditor getAuditor()
     {
         return auditor;
+    }
+
+    @VisibleForTesting
+    public void setBoundValueSuppressor(BoundValueSuppressor suppressor)
+    {
+        this.boundValueSuppressor = suppressor;
+    }
+
+    @VisibleForTesting
+    public BoundValueSuppressor getBoundValueSuppressor()
+    {
+        return boundValueSuppressor;
     }
 }